requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.7.0
asgiref==3.8.1
attrs==24.3.0
bitarray==3.0.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
ckzg==2.0.1
click==8.1.8
colorama==0.4.6
cytoolz==1.0.1
dataclasses-json==0.6.7
diskcache==5.6.3
distro==1.9.0
docstring_parser==0.16
elevenlabs==1.50.3
eth-account==0.13.4
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.6.1
eth-rlp==2.1.0
eth-typing==5.1.0
eth-utils==5.1.0
eth_abi==5.2.0
filelock==3.16.1
Flask==3.1.0
Flask-Cors==5.0.0
frozenlist==1.5.0
fsspec==2024.12.0
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-aiplatform==1.75.0
google-cloud-bigquery==3.27.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.14.0
google-cloud-storage==2.19.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
greenlet==3.1.1
groq==0.15.0
grpc-google-iam-v1==0.14.0
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
hexbytes==1.3.0
httpcore==1.0.7
httpx==0.27.2
httpx-sse==0.4.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.14
langchain-community==0.3.14
langchain-core==0.3.29
langchain-google-vertexai==2.0.9
langchain-groq==0.2.3
langchain-ollama==0.2.2
langchain-text-splitters==0.3.4
langsmith==0.2.10
llama_cpp_python==0.3.5
llvmlite==0.43.0
MarkupSafe==3.0.2
marshmallow==3.23.3
more-itertools==10.5.0
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
networkx==3.4.2
numba==0.60.0
numpy==1.26.4
ollama==0.4.5
omegaconf==2.3.0
openai-whisper==20240930
orjson==3.10.13
packaging==24.2
parsimonious==0.10.0
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.9.2
pydantic-settings==2.7.1
pydantic_core==2.23.4
pydub==0.25.1
pynput==1.7.7
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyunormalize==16.0.0
pywin32==308
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rlp==4.0.1
rsa==4.9
shapely==2.0.6
silero==0.4.1
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.13.0
SQLAlchemy==2.0.36
sympy==1.13.1
tenacity==9.0.0
tiktoken==0.8.0
toolz==1.0.0
torch==2.5.1
torchaudio==2.5.1
tqdm==4.67.1
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.3.0
Wave==0.0.2
web3==7.7.0
websockets==13.1
Werkzeug==3.1.3
whisper==1.1.10
yarl==1.18.3
